+++ projects/jrng-ui/tour/public-api.ts
@@ -1,0 +1,14 @@
+export { JTourService } from 'jrng-ui';
+export { JTourStepDirective } from 'jrng-ui';
+export type {
+  JTourAlign,
+  JTourButton,
+  JTourConfig,
+  JTourDefaults,
+  JTourEvent,
+  JTourEventType,
+  JTourSide,
+  JTourStep,
+  JTourStepInput,
+} from 'jrng-ui';
+